+++ CHANGELOG.md
@@ -8,9 +8,10 @@ Version latest (0.11.1rc?, 2019-09-24?)
     - this makes the code much simpler with `listen_sqs` call instead of all the internals showing up in `tagsSuggestAdvanced`
 - feat: http requests to aws api gateway to be sigv4-signed with aws keys (in order to do clean authentication on http)
     - this alleviates the need to do further authentication via sqs for example (where sqs is limited to the source account by iam policy)
-    - WIP
 - bugfix: redis cache to check *all* variables set
 - enh: when n ec2 > 10, prompt user to use redis caching
+- bugfix: http/sqs combo request was dropping messages since `dt_now` was calculated *after* request .. fixed
+- bugfix: sigv4 with aws-requests-auth updated
 
 
 Version 0.11.0 (2019-10-15)